HADOOP-17194. Splitting the builder and context class.
